+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Correct install of glibc encrypt on Gentoo.
+
 ## 0.5.0, release 2023-12
 * Update Voidlinux url https://repo-default.voidlinux.org.
 * Add global use="modules-sign" for Gentoo.